# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "manage-api-service",
        "description": "管理平台后端服务",
        "contact": {
            "name": "Rytia",
            "email": "admin@zzfly.net"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/get_template_list": {
            "get": {
                "tags": [
                    "template"
                ],
                "summary": "获取模板列表",
                "operationId": "App\\Http\\Controllers\\TemplateController::getList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "查询页码"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页返回条数"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "image": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "id": "1",
                                            "image": "http://www.baidu.com",
                                            "description": "测试简介"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/create_template": {
            "post": {
                "tags": [
                    "template"
                ],
                "summary": "新增模板",
                "operationId": "App\\Http\\Controllers\\TemplateController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image",
                                    "description"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "image": "http://www.baidu.com",
                                    "description": "测试简介"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/update_template": {
            "post": {
                "tags": [
                    "template"
                ],
                "summary": "更新模板",
                "description": "若此接口不传入 id，则与 create_template 行为一致(即以传入的信息新增模板)",
                "operationId": "App\\Http\\Controllers\\TemplateController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 6,
                                    "image": "http://www.baidu.com",
                                    "description": "测试简介"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/delete_template": {
            "post": {
                "tags": [
                    "template"
                ],
                "summary": "删除模板",
                "operationId": "App\\Http\\Controllers\\TemplateController::delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "operationId": "App\\Http\\Controllers\\UserController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@zzfly.net",
                                    "password": "520"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "token": "MSwxNTg4NDAzNzQ5LGNjYjcwZDc0MDJjN2EzNGYyYjE0YmNjNDEzYTU4NzUz"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "用户注销",
                "description": "后端无状态，前端取消 token 的使用即可",
                "operationId": "App\\Http\\Controllers\\UserController::logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "token": "MSwxNTg4NDAzNzQ5LGNjYjcwZDc0MDJjN2EzNGYyYjE0YmNjNDEzYTU4NzUz"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/get_user_list": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "获取用户列表",
                "operationId": "App\\Http\\Controllers\\UserController::getList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "查询页码"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "每页返回条数"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "example": {
                                            "id": 1,
                                            "email": "admin@zzfly.net",
                                            "password": "603794e417b556ab8c4a169eec84fa88"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/create_user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "新增用户",
                "operationId": "App\\Http\\Controllers\\UserController::create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@zzfly.net",
                                    "password": "520"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/update_user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "更新用户",
                "description": "若此接口不传入 id，则与 create_user 行为一致(即以传入的 email 和 password 新增用户)",
                "operationId": "App\\Http\\Controllers\\UserController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 9,
                                    "email": "admin@zzfly.net",
                                    "password": "520"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/delete_user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "operationId": "App\\Http\\Controllers\\UserController::delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 9
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}